+++ scripts/components/result/details/homology.jsx
@@ -1,38 +1,27 @@
-'use strict';
+import React from "react";
+import {Detail, Title, Description, Content, Explore, Links} from "./generic/detail.jsx";
+import TreeVis from "gramene-genetree-vis";
+import queryActions from "../../../actions/queryActions";
+import DocActions from "../../../actions/docActions";
+import lutStore from "../../../stores/lutStore";
+import _ from "lodash";
+import treesClient from "gramene-trees-client";
+const processGenetreeDoc = treesClient.genetree.tree;
 
-var React = require('react');
-var Reflux = require('reflux');
+export default class Homology extends React.Component {
 
-var TreeVis = require('gramene-genetree-vis');
-
-var queryActions = require('../../../actions/queryActions');
-var DocActions = require('../../../actions/docActions');
-var lutStore = require('../../../stores/lutStore');
-
-var _ = require('lodash');
-
-var processGenetreeDoc = require('gramene-trees-client').genetree.tree;
-
-var QueryTerm = require('../queryTerm.jsx');
-
-var Homology = React.createClass({
-
-  propTypes: {
-    gene: React.PropTypes.object.isRequired,
-    docs: React.PropTypes.object.isRequired
-  },
-
-  mixins: [
-    Reflux.connect(lutStore, 'luts')
-  ],
-
-  getInitialState: function () {
-    return {
-      luts: lutStore.state
+  constructor(props) {
+    super(props);
+    this.state = {
+      taxonomy: _.get(lutStore, 'state.taxonomy')
     };
-  },
+  }
 
-  componentWillMount: function () {
+  componentWillMount() {
+    this.unsubscribe = lutStore.listen((lutState) =>
+      this.setState({taxonomy: lutState.taxonomy})
+    );
+
     this.orthologs = this.orthologList();
     this.paralogs = this.paralogList();
     this.treeId = _.get(this.props.gene, 'homology.gene_tree.id');
@@ -40,26 +29,26 @@ var Homology = React.createClass({
     if (this.treeId) {
       DocActions.needDocs('genetrees', this.treeId, processGenetreeDoc);
     }
-  },
+  }
 
-  componentWillUpdate: function (newProps, newState) {
-    this.genetree = _.get(newProps, ['docs', 'genetrees', this.treeId]);
-    this.taxonomy = _.get(newState, ['luts', 'taxonomy']);
-  },
-
-  componentWillUnmount: function () {
+  componentWillUnmount() {
+    this.unsubscribe();
     DocActions.noLongerNeedDocs('genetrees', this.treeId);
-  },
+  }
 
-  orthologList: function () {
+  componentWillUpdate(newProps) {
+    this.genetree = _.get(newProps, ['docs', 'genetrees', this.treeId]);
+  }
+
+  orthologList() {
     return this.orthoParaList('ortholog');
-  },
+  }
 
-  paralogList: function () {
+  paralogList() {
     return this.orthoParaList('within_species_paralog');
-  },
+  }
 
-  orthoParaList: function orthoParaList(type) {
+  orthoParaList(type) {
     var homology, thisGeneId;
     homology = _.get(this.props.gene, 'homology.homologous_genes');
     thisGeneId = this.props.gene._id;
@@ -78,9 +67,9 @@ var Homology = React.createClass({
         return homologs; // only return something if we have something. We're testing for truthiness later.
       }
     }
-  },
+  }
 
-  createAllHomologsFilters: function () {
+  createAllHomologsFilters() {
     var gt, fq, result;
     gt = this.treeId;
     fq = 'gene_tree:' + gt;
@@ -92,9 +81,9 @@ var Homology = React.createClass({
       display_name: 'Homolog of ' + this.props.gene.name
     };
     return result;
-  },
+  }
 
-  createOrthologFilters: function () {
+  createOrthologFilters() {
     var fq, id, result;
     id = this.props.gene._id;
     fq = 'homology__ortholog_one2one:' + id +
@@ -109,9 +98,9 @@ var Homology = React.createClass({
       display_name: 'Orthologs of ' + this.props.gene.name
     };
     return result;
-  },
+  }
 
-  createParalogFilters: function () {
+  createParalogFilters() {
     var fq, id, result;
     id = this.props.gene._id;
     fq = 'homology__within_species_paralog:' + id +
@@ -124,36 +113,36 @@ var Homology = React.createClass({
       display_name: 'Paralogs of ' + this.props.gene.name
     };
     return result;
-  },
+  }
 
-  filterAllHomologs: function () {
+  filterAllHomologs() {
     queryActions.setAllFilters(this.createAllHomologsFilters());
-  },
+  }
 
-  filterOrthologs: function () {
+  filterOrthologs() {
     queryActions.setAllFilters(this.createOrthologFilters());
-  },
+  }
 
-  filterParalogs: function () {
+  filterParalogs() {
     queryActions.setAllFilters(this.createParalogFilters());
-  },
+  }
 
-  renderTreeVis: function () {
-    if (this.genetree && this.taxonomy) {
+  renderTreeVis() {
+    if (this.genetree && this.state.taxonomy) {
       return (
         <div className="gene-genetree">
           <h5>Gene Tree</h5>
           <TreeVis genetree={this.genetree}
                    initialGeneOfInterest={this.props.gene}
-                   taxonomy={this.taxonomy}
+                   taxonomy={this.state.taxonomy}
                    width={600}/>
         </div>
       );
     }
-  },
+  }
 
-  renderFilters: function () {
-    var tree, geneCount, filters;
+  explorations() {
+    var tree, geneCount, explorations;
 
     if (this.treeId) {
       tree = _.get(this.props, ['docs', 'genetrees', this.treeId]);
@@ -163,62 +152,58 @@ var Homology = React.createClass({
       geneCount = tree.geneCount;
     }
 
-    filters = [
-      <QueryTerm key="all"
-                 name="Show All Homologs"
-                 count={geneCount}
-                 handleClick={this.filterAllHomologs}/>
+    explorations = [
+      {
+        name: 'Show All Homologs',
+        handleClick: this.filterAllHomologs.bind(this),
+        count: geneCount
+      }
     ];
+
     if (this.orthologs) {
-      filters.push(<QueryTerm key="ortho"
-                              name="Show Orthologs"
-                              count={this.orthologs.length}
-                              handleClick={this.filterOrthologs}/>)
+      explorations.push({
+        name: "Show Orthologs",
+        handleClick: this.filterOrthologs.bind(this),
+        count: this.orthologs.length
+      });
     }
     if (this.paralogs) {
-      filters.push(<QueryTerm key="para"
-                              name="Show Paralogs"
-                              count={this.paralogs.length}
-                              handleClick={this.filterParalogs}/>)
+      explorations.push({
+        name: "Show Paralogs",
+        handleClick: this.filterParalogs.bind(this),
+        count: this.paralogs.length
+      });
     }
 
-    return (
-      <div>
-        <h5>Change the query</h5>
-        <ul>
-          {filters}
-        </ul>
-      </div>
-    );
-  },
+    return explorations;
+  }
 
-  renderLinks: function () {
+  links() {
     var gene, ensemblGenetreeUrl;
 
     gene = this.props.gene;
     ensemblGenetreeUrl = '//ensembl.gramene.org/' + gene.system_name + '/Gene/Compara_Tree?g=' + gene._id;
 
-    return (
-      <div>
-        <h5>Links</h5>
-        <ul>
-          <li>
-            <a href={ensemblGenetreeUrl}>Ensembl Gene Tree view</a>
-          </li>
-        </ul>
-      </div>
-    )
-  },
+    return [
+      {name: 'Ensembl Gene Tree view', url: ensemblGenetreeUrl}
+    ];
+  }
 
-  render: function () {
+  render() {
     return (
-      <div>
-        {this.renderTreeVis()}
-        {this.renderFilters()}
-        {this.renderLinks()}
-      </div>
+      <Detail>
+        <Title>Compara Gene Tree</Title>
+        <Description>This phylogram shows the relationships between this genes and others similar to it, as determined
+          by <a target="_blank" href="http://ensembl.org/info/genome/compara/index.html">Ensembl Compara</a>.</Description>
+        <Content>{this.renderTreeVis()}</Content>
+        <Explore explorations={this.explorations()}/>
+        <Links links={this.links()}/>
+      </Detail>
     );
   }
-});
+}
 
-module.exports = Homology;
+Homology.propTypes = {
+  gene: React.PropTypes.object.isRequired,
+  docs: React.PropTypes.object.isRequired
+};
